requirements.txt
asgiref==3.9.1
cachetools==6.2.0
Django==5.2.6
django-jazzmin==3.0.1
django-rql==4.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
lark-parser==0.11.0
lib-rql==2.0.2
mccabe==0.7.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
